ui: Use `element-plus` instead of `naive-ui`
When some components need custom rendering, slot cannot be used, only the
h function can be used in `naive-ui`. As a typical example, when using a
table component, you often need to do a custom rendering of a column of
data returned by the background. At this point in the `.vue` single file
can only use the h function to perform custom rendering.

Signed-off-by: Jianhui Zhao <zhaojh329@gmail.com>